✅ resolving dbt run errors

+++ dbt_project/models/kids_first/chr_18/src/S3_Bucket_Scrape/kf_chr_18_src_s3_scrape_cody.sql
@@ -2,7 +2,7 @@
 
     select
     ROW_NUMBER() OVER () AS kf_index,
-    "Size"::integer as size,
+    "Size"::bigint as size,
     "Key"::text as key,
     "ETag"::text as etag,
     "Bucket"::text as bucket,
